pyproject.toml
[project]
name = "claude-controltower"
version = "0.1.0"
description = "Claude Code Orchestration System - 텔레그램을 통한 Claude Code 원격 제어"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiosqlite>=0.20.0",
    "structlog>=24.0",
    "psutil>=6.0",
    "mss>=9.0",
    "pystray>=0.19",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov",
    "ruff",
    "mypy",
]
build = [
    "pyinstaller>=6.0",
]

[project.scripts]
controltower = "src.main:entry"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
